(BSR)[API] refactor: move collective functions from offer to educational repository
