+++ api/src/pcapi/core/offers/repository.py
@@ -6,15 +6,12 @@ import typing
 
 import pytz
 import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
 import sqlalchemy.exc as sa_exc
 import sqlalchemy.orm as sa_orm
 from sqlalchemy.sql import and_
-from sqlalchemy.sql import or_
 
 from pcapi.core.bookings import models as bookings_models
 from pcapi.core.categories import subcategories
-from pcapi.core.categories.models import EacFormat
 from pcapi.core.educational import models as educational_models
 from pcapi.core.geography import models as geography_models
 from pcapi.core.offerers import models as offerers_models
@@ -452,163 +449,6 @@ def get_offers_by_filters(
     return query
 
 
-def get_collective_offers_by_filters(
-    *,
-    user_id: int,
-    user_is_admin: bool,
-    offerer_id: int | None = None,
-    statuses: list[educational_models.CollectiveOfferDisplayedStatus] | None = None,
-    venue_id: int | None = None,
-    provider_id: int | None = None,
-    name_keywords: str | None = None,
-    period_beginning_date: datetime.date | None = None,
-    period_ending_date: datetime.date | None = None,
-    formats: list[EacFormat] | None = None,
-) -> BaseQuery:
-    query = db.session.query(educational_models.CollectiveOffer)
-
-    if not user_is_admin:
-        query = (
-            query.join(offerers_models.Venue)
-            .join(offerers_models.Offerer)
-            .join(offerers_models.UserOfferer)
-            .filter(
-                offerers_models.UserOfferer.userId == user_id,
-                offerers_models.UserOfferer.isValidated,
-            )
-        )
-
-    if offerer_id is not None:
-        if user_is_admin:
-            query = query.join(offerers_models.Venue)
-        query = query.filter(offerers_models.Venue.managingOffererId == offerer_id)
-
-    if venue_id is not None:
-        query = query.filter(educational_models.CollectiveOffer.venueId == venue_id)
-
-    if provider_id is not None:
-        query = query.filter(educational_models.CollectiveOffer.providerId == provider_id)
-
-    if name_keywords is not None:
-        search = name_keywords
-        if len(name_keywords) > 3:
-            search = "%{}%".format(name_keywords)
-        # We should really be using `union` instead of `union_all` here since we don't want duplicates but
-        # 1. it's unlikely that a book will contain its EAN in its name
-        # 2. we need to migrate models.Offer.extraData to JSONB in order to use `union`
-        query = query.filter(educational_models.CollectiveOffer.name.ilike(search))
-
-    if statuses:
-        query = _filter_collective_offers_by_statuses(query, statuses)
-
-    if period_beginning_date is not None or period_ending_date is not None:
-        subquery = (
-            db.session.query(educational_models.CollectiveStock)
-            .with_entities(educational_models.CollectiveStock.collectiveOfferId)
-            .distinct(educational_models.CollectiveStock.collectiveOfferId)
-            .join(educational_models.CollectiveOffer)
-            .join(offerers_models.Venue)
-        )
-        if period_beginning_date is not None:
-            subquery = subquery.filter(
-                sa.func.timezone(
-                    offerers_models.Venue.timezone,
-                    sa.func.timezone("UTC", educational_models.CollectiveStock.startDatetime),
-                )
-                >= period_beginning_date
-            )
-        if period_ending_date is not None:
-            subquery = subquery.filter(
-                sa.func.timezone(
-                    offerers_models.Venue.timezone,
-                    sa.func.timezone("UTC", educational_models.CollectiveStock.startDatetime),
-                )
-                <= datetime.datetime.combine(period_ending_date, datetime.time.max),
-            )
-        if venue_id is not None:
-            subquery = subquery.filter(educational_models.CollectiveOffer.venueId == venue_id)
-        elif offerer_id is not None:
-            subquery = subquery.filter(offerers_models.Venue.managingOffererId == offerer_id)
-        elif not user_is_admin:
-            subquery = (
-                subquery.join(offerers_models.Offerer)
-                .join(offerers_models.UserOfferer)
-                .filter(
-                    offerers_models.UserOfferer.userId == user_id,
-                    offerers_models.UserOfferer.isValidated,
-                )
-            )
-        q2 = subquery.subquery()
-        query = query.join(q2, q2.c.collectiveOfferId == educational_models.CollectiveOffer.id)
-
-    if formats:
-        query = query.filter(
-            educational_models.CollectiveOffer.formats.overlap(postgresql.array((format.name for format in formats)))
-        )
-
-    return query
-
-
-def get_collective_offers_template_by_filters(
-    *,
-    user_id: int,
-    user_is_admin: bool,
-    offerer_id: int | None = None,
-    statuses: list[educational_models.CollectiveOfferDisplayedStatus] | None = None,
-    venue_id: int | None = None,
-    name_keywords: str | None = None,
-    period_beginning_date: datetime.date | None = None,
-    period_ending_date: datetime.date | None = None,
-    formats: list[EacFormat] | None = None,
-) -> BaseQuery:
-    query = db.session.query(educational_models.CollectiveOfferTemplate)
-
-    if period_beginning_date is not None or period_ending_date is not None:
-        query = query.filter(sa.false())
-
-    if not user_is_admin:
-        query = (
-            query.join(offerers_models.Venue)
-            .join(offerers_models.Offerer)
-            .join(offerers_models.UserOfferer)
-            .filter(
-                offerers_models.UserOfferer.userId == user_id,
-                offerers_models.UserOfferer.isValidated,
-            )
-        )
-
-    if offerer_id is not None:
-        if user_is_admin:
-            query = query.join(offerers_models.Venue)
-        query = query.filter(offerers_models.Venue.managingOffererId == offerer_id)
-
-    if venue_id is not None:
-        query = query.filter(educational_models.CollectiveOfferTemplate.venueId == venue_id)
-
-    if name_keywords is not None:
-        search = name_keywords
-        if len(name_keywords) > 3:
-            search = "%{}%".format(name_keywords)
-        # We should really be using `union` instead of `union_all` here since we don't want duplicates but
-        # 1. it's unlikely that a book will contain its EAN in its name
-        # 2. we need to migrate models.Offer.extraData to JSONB in order to use `union`
-        query = query.filter(educational_models.CollectiveOfferTemplate.name.ilike(search))
-
-    if statuses:
-        template_statuses = set(statuses) & set(educational_models.COLLECTIVE_OFFER_TEMPLATE_STATUSES)
-        status_values = [status.value for status in template_statuses]
-        query = query.filter(educational_models.CollectiveOfferTemplate.displayedStatus.in_(status_values))  # type: ignore[attr-defined]
-
-    if formats:
-        query = query.filter(
-            educational_models.CollectiveOfferTemplate.formats.overlap(
-                postgresql.array((format.name for format in formats))
-            )
-        )
-
-    return query
-
-
 def _filter_by_creation_mode(query: BaseQuery, creation_mode: str) -> BaseQuery:
     if creation_mode == MANUAL_CREATION_MODE:
         query = query.filter(models.Offer.lastProviderId.is_(None))
@@ -620,234 +460,6 @@ def _filter_by_creation_mode(query: BaseQuery, creation_mode: str) -> BaseQuery:
 
 def _filter_by_status(query: BaseQuery, status: str) -> BaseQuery:
     return query.filter(models.Offer.status == offer_mixin.OfferStatus[status].name)
-
-
-def _filter_collective_offers_by_statuses(
-    query: BaseQuery,
-    statuses: list[educational_models.CollectiveOfferDisplayedStatus] | None,
-) -> BaseQuery:
-    """
-    Filter a SQLAlchemy query for CollectiveOffers based on a list of statuses.
-
-    This function modifies the input query to filter CollectiveOffers based on their CollectiveOfferDisplayedStatus.
-
-    Args:
-      query (BaseQuery): The initial query to be filtered.
-      statuses (list[CollectiveOfferDisplayedStatus]): A list of status strings to filter by.
-
-    Returns:
-      BaseQuery: The modified query with applied filters.
-    """
-    on_collective_offer_filters: list = []
-    on_booking_status_filter: list = []
-
-    if not statuses:
-        # if statuses is empty we return all offers
-        return query
-
-    offer_id_with_booking_status_subquery, query_with_booking = add_last_booking_status_to_collective_offer_query(query)
-
-    if educational_models.CollectiveOfferDisplayedStatus.ARCHIVED in statuses:
-        on_collective_offer_filters.append(educational_models.CollectiveOffer.isArchived == True)
-
-    if educational_models.CollectiveOfferDisplayedStatus.DRAFT in statuses:
-        on_collective_offer_filters.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.DRAFT,
-                educational_models.CollectiveOffer.isArchived == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.UNDER_REVIEW in statuses:
-        on_collective_offer_filters.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.PENDING,
-                educational_models.CollectiveOffer.isArchived == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.REJECTED in statuses:
-        on_collective_offer_filters.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.REJECTED,
-                educational_models.CollectiveOffer.isArchived == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.HIDDEN in statuses:
-        # if the statuses filter contains HIDDEN only, we need to return no collective_offer
-        # otherwise we return offers depending on the other statuses in the filter
-        on_collective_offer_filters.append(sa.false())
-
-    if educational_models.CollectiveOfferDisplayedStatus.PUBLISHED in statuses:
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == None,
-                educational_models.CollectiveOffer.hasBookingLimitDatetimesPassed == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.PREBOOKED in statuses:
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.PENDING,
-                educational_models.CollectiveOffer.hasBookingLimitDatetimesPassed == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.BOOKED in statuses:
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.CONFIRMED,
-                educational_models.CollectiveOffer.hasEndDatetimePassed == False,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.ENDED in statuses:
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                or_(
-                    offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.USED,
-                    offer_id_with_booking_status_subquery.c.status
-                    == educational_models.CollectiveBookingStatus.CONFIRMED,
-                ),
-                educational_models.CollectiveOffer.hasEndDatetimePassed == True,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.REIMBURSED in statuses:
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.REIMBURSED,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.EXPIRED in statuses:
-        # expired with a pending booking or no booking
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                educational_models.CollectiveOffer.hasBookingLimitDatetimesPassed == True,
-                educational_models.CollectiveOffer.hasStartDatetimePassed == False,
-                or_(
-                    offer_id_with_booking_status_subquery.c.status
-                    == educational_models.CollectiveBookingStatus.PENDING,
-                    offer_id_with_booking_status_subquery.c.status == None,
-                ),
-            )
-        )
-        # expired with a booking cancelled with reason EXPIRED
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                educational_models.CollectiveOffer.hasBookingLimitDatetimesPassed == True,
-                educational_models.CollectiveOffer.hasStartDatetimePassed == False,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.CANCELLED,
-                offer_id_with_booking_status_subquery.c.cancellationReason
-                == educational_models.CollectiveBookingCancellationReasons.EXPIRED,
-            )
-        )
-
-    if educational_models.CollectiveOfferDisplayedStatus.CANCELLED in statuses:
-        # Cancelled due to expired booking
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.CANCELLED,
-                offer_id_with_booking_status_subquery.c.cancellationReason
-                == educational_models.CollectiveBookingCancellationReasons.EXPIRED,
-                educational_models.CollectiveOffer.hasStartDatetimePassed == True,
-            )
-        )
-
-        # Cancelled by admin / CA or on ADAGE
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == educational_models.CollectiveBookingStatus.CANCELLED,
-                offer_id_with_booking_status_subquery.c.cancellationReason
-                != educational_models.CollectiveBookingCancellationReasons.EXPIRED,
-            ),
-        )
-
-        # Cancelled due to no booking when the event has started
-        on_booking_status_filter.append(
-            and_(
-                educational_models.CollectiveOffer.validation == offer_mixin.OfferValidationStatus.APPROVED,
-                educational_models.CollectiveOffer.isActive == True,
-                offer_id_with_booking_status_subquery.c.status == None,
-                educational_models.CollectiveOffer.hasStartDatetimePassed == True,
-            ),
-        )
-
-    # Add filters on `CollectiveBooking.Status`
-    if on_booking_status_filter:
-        substmt = query_with_booking.filter(or_(*on_booking_status_filter)).subquery()
-        on_collective_offer_filters.append(educational_models.CollectiveOffer.id.in_(sa.select(substmt.c.id)))
-
-    # Add filters on `CollectiveOffer`
-    if on_collective_offer_filters:
-        query = query.filter(or_(*on_collective_offer_filters))
-
-    return query
-
-
-def add_last_booking_status_to_collective_offer_query(
-    query: BaseQuery,
-) -> typing.Tuple[typing.Any, BaseQuery]:
-    last_booking_query = (
-        db.session.query(educational_models.CollectiveBooking)
-        .with_entities(
-            educational_models.CollectiveBooking.collectiveStockId,
-            sa.func.max(educational_models.CollectiveBooking.dateCreated).label("maxdate"),
-        )
-        .group_by(educational_models.CollectiveBooking.collectiveStockId)
-        .subquery()
-    )
-
-    collective_stock_with_last_booking_status_query = (
-        db.session.query(educational_models.CollectiveStock)
-        .with_entities(
-            educational_models.CollectiveStock.collectiveOfferId,
-            educational_models.CollectiveStock.bookingLimitDatetime,
-            educational_models.CollectiveBooking.status,
-            educational_models.CollectiveBooking.cancellationReason,
-        )
-        .outerjoin(
-            educational_models.CollectiveBooking,
-            educational_models.CollectiveStock.collectiveBookings,
-        )
-        .join(
-            last_booking_query,
-            sa.and_(
-                educational_models.CollectiveBooking.collectiveStockId == last_booking_query.c.collectiveStockId,
-                educational_models.CollectiveBooking.dateCreated == last_booking_query.c.maxdate,
-            ),
-        )
-    )
-
-    subquery = collective_stock_with_last_booking_status_query.subquery()
-
-    query_with_booking = query.outerjoin(
-        subquery,
-        subquery.c.collectiveOfferId == educational_models.CollectiveOffer.id,
-    )
-
-    return subquery, query_with_booking
 
 
 def get_products_map_by_provider_reference(
