+++ api/tests/core/offers/test_repository.py
@@ -11,9 +11,6 @@ import time_machine
 import pcapi.core.bookings.factories as bookings_factories
 from pcapi.core.categories import pro_categories
 from pcapi.core.categories import subcategories
-from pcapi.core.categories.models import EacFormat
-import pcapi.core.educational.factories as educational_factories
-import pcapi.core.educational.models as educational_models
 from pcapi.core.geography import factories as geography_factories
 import pcapi.core.offerers.factories as offerers_factories
 from pcapi.core.offers import exceptions
@@ -1421,40 +1418,6 @@ class AvailableActivationCodeTest:
         assert not repository.get_available_activation_code(stock)
 
 
-@pytest.fixture(name="admin_user", scope="class")
-def admin_user_fixture():
-    return users_factories.AdminFactory()
-
-
-@pytest.mark.usefixtures("db_session")
-class GetCollectiveOffersTemplateByFiltersTest:
-    @pytest.mark.parametrize("offer_status", educational_models.COLLECTIVE_OFFER_TEMPLATE_STATUSES)
-    def test_filter_each_status(self, admin_user, offer_status):
-        template = educational_factories.create_collective_offer_template_by_status(offer_status)
-
-        result = repository.get_collective_offers_template_by_filters(
-            user_id=admin_user.id, user_is_admin=True, statuses=[offer_status]
-        )
-        assert result.one() == template
-
-        statuses_no_result = set(educational_models.CollectiveOfferDisplayedStatus) - {offer_status}
-        result = repository.get_collective_offers_template_by_filters(
-            user_id=admin_user.id, user_is_admin=True, statuses=statuses_no_result
-        )
-        assert result.count() == 0
-
-    def test_formats_filter(self, admin_user):
-        template = educational_factories.CollectiveOfferTemplateFactory(
-            formats=[EacFormat.CONCERT, EacFormat.CONFERENCE_RENCONTRE]
-        )
-        educational_factories.CollectiveOfferTemplateFactory(formats=[EacFormat.CONFERENCE_RENCONTRE])
-
-        result = repository.get_collective_offers_template_by_filters(
-            user_id=admin_user.id, user_is_admin=True, formats=[EacFormat.CONCERT]
-        ).one()
-        assert result.id == template.id
-
-
 class UpdateStockQuantityToDnBookedQuantityTest:
     def test_update_stock_quantity_to_dn_booked_quantity(self, db_session):
         # given
@@ -1516,205 +1479,6 @@ class GetPaginatedOfferIdsByVenueIdTest:
         offer2 = factories.OfferFactory(venue=venue, isActive=False)
 
         assert repository.get_paginated_offer_ids_by_venue_id(venue.id, limit=2, page=0) == [offer1.id, offer2.id]
-
-
-@pytest.mark.usefixtures("db_session")
-class GetFilteredCollectiveOffersTest:
-    def test_get_prebooked_collective_offers(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-        collective_offer_prebooked = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer
-        )
-        collective_stock_prebooked = educational_factories.CollectiveStockFactory(
-            collectiveOffer=collective_offer_prebooked
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_prebooked,
-            status=educational_models.CollectiveBookingStatus.CANCELLED.value,
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_prebooked,
-            status=educational_models.CollectiveBookingStatus.PENDING.value,
-        )
-
-        collective_offer_cancelled = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer
-        )
-        collective_stock_cancelled = educational_factories.CollectiveStockFactory(
-            collectiveOffer=collective_offer_cancelled
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_cancelled,
-            status=educational_models.CollectiveBookingStatus.PENDING.value,
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_cancelled,
-            status=educational_models.CollectiveBookingStatus.CANCELLED.value,
-        )
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.PREBOOKED],
-        )
-        assert offers.all() == [collective_offer_prebooked]
-
-    def test_get_ended_collective_offers(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-        collective_offer_ended = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer
-        )
-        collective_stock_ended = educational_factories.CollectiveStockFactory(
-            collectiveOffer=collective_offer_ended,
-            startDatetime=datetime.datetime(year=2000, month=1, day=1),
-        )
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_ended,
-            status=educational_models.CollectiveBookingStatus.USED.value,
-        )
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.ENDED],
-        )
-        assert offers.all() == [collective_offer_ended]
-
-    def test_get_collective_offers_with_formats(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-        collective_offer = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer, formats=[EacFormat.CONCERT]
-        )
-        educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer,
-            formats=[EacFormat.CONFERENCE_RENCONTRE],
-        )
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            formats=[EacFormat.CONCERT],
-        )
-        assert offers.one() == collective_offer
-
-    def test_get_collective_offers_archived(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-
-        _collective_offer_unarchived = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer
-        )
-        collective_offer_archived = educational_factories.CollectiveOfferFactory(
-            venue__managingOfferer=user_offerer.offerer,
-            isActive=False,
-            dateArchived=datetime.datetime(year=2000, month=1, day=1),
-        )
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.ARCHIVED],
-        )
-        assert offers.one() == collective_offer_archived
-
-    def test_get_collective_offers_draft(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-
-        collective_offer_draft = educational_factories.CollectiveOfferFactory(
-            validation=offer_mixin.OfferValidationStatus.DRAFT.value,
-            venue__managingOfferer=user_offerer.offerer,
-        )
-
-        offers = repository.get_collective_offers_by_filters(user_id=user_offerer.userId, user_is_admin=False)
-        with assert_num_queries(1):
-            assert offers.one() == collective_offer_draft
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.DRAFT],
-        )
-        with assert_num_queries(1):
-            assert offers.one() == collective_offer_draft
-
-    def test_get_collective_offers_expired(self):
-        user_offerer = offerers_factories.UserOffererFactory()
-        now = datetime.datetime.utcnow()
-        future = now + datetime.timedelta(days=5)
-        past = now - datetime.timedelta(days=2)
-
-        # expired offer without booking
-        collective_offer_expired = educational_factories.CollectiveOfferFactory(
-            validation=offer_mixin.OfferValidationStatus.APPROVED.value,
-            venue__managingOfferer=user_offerer.offerer,
-        )
-
-        educational_factories.CollectiveStockFactory(
-            collectiveOffer=collective_offer_expired,
-            startDatetime=future,
-            bookingLimitDatetime=past,
-        )
-
-        # expired offer with pending booking
-        collective_offer_prebooked_expired = educational_factories.CollectiveOfferFactory(
-            validation=offer_mixin.OfferValidationStatus.APPROVED.value,
-            venue__managingOfferer=user_offerer.offerer,
-        )
-
-        collective_stock_prebooked_expired = educational_factories.CollectiveStockFactory(
-            collectiveOffer=collective_offer_prebooked_expired,
-            startDatetime=future,
-            bookingLimitDatetime=past,
-        )
-
-        educational_factories.CollectiveBookingFactory(
-            collectiveStock=collective_stock_prebooked_expired,
-            confirmationLimitDate=past,
-            status=educational_models.CollectiveBookingStatus.CANCELLED.value,
-            cancellationReason=educational_models.CollectiveBookingCancellationReasons.EXPIRED.value,
-        )
-
-        offers = repository.get_collective_offers_by_filters(
-            user_id=user_offerer.userId,
-            user_is_admin=False,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.EXPIRED],
-        )
-        offers_list = offers.all()
-        assert len(offers_list) == 2
-        assert set(offers_list) == {
-            collective_offer_expired,
-            collective_offer_prebooked_expired,
-        }
-
-    @pytest.mark.parametrize(
-        "offer_status",
-        set(educational_models.CollectiveOfferDisplayedStatus)
-        - {educational_models.CollectiveOfferDisplayedStatus.HIDDEN},
-    )
-    def test_filter_each_status(self, admin_user, offer_status):
-        offer = educational_factories.create_collective_offer_by_status(offer_status)
-
-        result = repository.get_collective_offers_by_filters(
-            user_id=admin_user.id, user_is_admin=True, statuses=[offer_status]
-        )
-        assert result.one() == offer
-
-        statuses_no_result = set(educational_models.CollectiveOfferDisplayedStatus) - {offer_status}
-        result = repository.get_collective_offers_by_filters(
-            user_id=admin_user.id, user_is_admin=True, statuses=statuses_no_result
-        )
-        assert result.count() == 0
-
-    def test_filter_hidden(self, admin_user):
-        for status in educational_models.CollectiveOfferDisplayedStatus:
-            educational_factories.create_collective_offer_by_status(status)
-
-        # The HIDDEN filter does not correspond to any collective offer (only templates)
-        result = repository.get_collective_offers_by_filters(
-            user_id=admin_user.id,
-            user_is_admin=True,
-            statuses=[educational_models.CollectiveOfferDisplayedStatus.HIDDEN],
-        )
-        assert result.count() == 0
 
 
 @pytest.mark.usefixtures("db_session")
